requirements.txt
discord.py
bs4
Mako
ruamel.yaml